benches: add a network benchmark
