+++ benchmarks/Cargo.toml
@@ -12,6 +12,11 @@ license = "GPL-3.0"
 edition = "2018"
 
 [[bench]]
+name = "network"
+path = "network/network.rs"
+harness = false
+
+[[bench]]
 name = "posw"
 path = "posw/posw.rs"
 harness = false
@@ -24,18 +29,15 @@ snarkvm-utilities = { git = "https://github.com/AleoHQ/snarkVM.git", rev = "a7e0
 
 snarkos-profiler = { path = "../profiler", version = "1.1.4" }
 snarkos-posw = { path = "../posw", version = "1.1.4", features = ["test-helpers"] }
+snarkos-network = { path = "../network" }
+snarkos-testing = { path = "../testing" }
 
+csv = { version = "1" }
 derivative = { version = "2" }
 digest = { version = "0.7" }
 rand = { version = "0.7" }
 rayon = { version = "1" }
 smallvec = { version = "1.6" }
-criterion = "0.3.3"
-
-[dev-dependencies]
-snarkvm-curves = { git = "https://github.com/AleoHQ/snarkVM.git", rev = "a7e017b", version = "0.0.2"}
-
-criterion = "0.3.3"
-csv = { version = "1" }
+criterion = { version = "0.3.4", features = ["async_tokio"] }
 rand_xorshift = { version = "0.2" }
 tokio = { version = "1", features = ["full"] }
