+++ benchmarks/network/network.rs
@@ -1,0 +1,48 @@
+// Copyright (C) 2019-2020 Aleo Systems Inc.
+// This file is part of the snarkOS library.
+
+// The snarkOS library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkOS library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
+
+use criterion::*;
+use rand::{distributions::Standard, thread_rng, Rng};
+
+use snarkos_network::external::Payload;
+use snarkos_testing::network::spawn_2_fake_nodes;
+
+fn send_small_messages(c: &mut Criterion) {
+    let rt = tokio::runtime::Runtime::new().unwrap();
+
+    let (node0, node1) = rt.block_on(spawn_2_fake_nodes());
+    let node0 = tokio::sync::Mutex::new(node0);
+    let node1 = tokio::sync::Mutex::new(node1);
+
+    c.bench_function("send_small_messages", move |b| {
+        b.to_async(&rt).iter(|| async {
+            let block_size: u16 = thread_rng().gen();
+            let fake_block_bytes: Vec<u8> = (&mut thread_rng())
+                .sample_iter(Standard)
+                .take(block_size as usize)
+                .collect();
+            let big_block = Payload::Block(fake_block_bytes);
+
+            // send it from node0 to node1
+            node0.lock().await.write_message(&big_block).await;
+            let _payload = node1.lock().await.read_payload().await.unwrap();
+        })
+    });
+}
+
+criterion_group!(send_receive_benches, send_small_messages);
+
+criterion_main!(send_receive_benches);
